replace default hashheader parser
the default parser for headings in markdown accepts input like `#i am a
heading` and comes into conflict with our tag reference parser. this fix
slightly alters the original hashheader regex and requires that the user
places one or more spaces between the hash and the actual title of the
heading so only input like `# i am a heading` is valid.