fix retrieval of gestalt profile data. refs #596
The dom-select adapter used for the author-selection in articles queries
our gestalt rest api. unfortunately it used the get method of the
gestalt adapter as if it were the list method and later filtered the
gestalt object from that. the new implementation just fetches the
gestalt based on the get method.