disable forced format conversions for images
Django-Imagekit is smart enough to detect the format from the source
file. One (any maybe the only) reason we forced PNG as format is that
our previous image conversion tool could not auto-detect the format but
we still wanted PNG images with transparency to be handled…
transparently :).

A nasty side-effect of the forced conversion is that JPEG images were
converted to PNG which sometimes increased the file size of these images
multiple orders of magnitude. Often a large portion of our page loading
times on the frontpage of stadtgestalten comes from these kind of
images. Format auto-detection should fix this.