database_dump: avoid pg_dump warning about permissions
pg_dump is switching to a different directory and then tries to return
to the original directory during its runtime. This can lead to the
emission of a (harmless) warning, that it failed to change the directory
(since the script may have been called by root from a non-accessible
directory).
Now we switch to a neutral directory and take care to use absolute
paths.