Parse the "Date" field of incoming mails and apply it to the new post
Using the Date field instead of the arrival time is closer to the
behaviour of a mailing list.  But it obviously allows senders to inject
posts with a past or future timestamp.  This is not visible in the web
interface, thus we need to help our users to understand wrong
timestamps.  This issue to be addressed separately.