feat(events): allow management of event attendance for other group members
This allows a group to faciliate a simple kind of ressource management.
Users can create events/tasks and allocated themself or other members to
these events/tasks.
This feature can be enabled by a site-wide setting.

The feature is disabled by default.
The following configuration entry enables it:

```yaml
events:
  allow_attendance_management_for_group_members: true
```