Makefile: use "which" instead of "command -v"
"command -v" is the preferable and platform independent approach for
finding executables.  But it is a shell builtin, thus it is only usable,
if the command is executed in a shell and not executed directly (via
"execvpe").  The current version of GNU Make (v4.3) executes commands
with shell meta-characters (e.g. `||`) in a shell (`sh -c '...'`).
In contrast, all simple commands are executed directly.

Thus `$(shell command -v foo || true)` works.
But `$(shell command -v foo)` fails.

These implementation details of the "shell" function are not specified
for GNU Make.  Thus we should not rely on them and not force users to
memorize these edge cases.

Thus we should use `which` instead.