make: use site-wide installed packages by default
Previously the packages of the virtualenv were upgraded automatically.
This made it impossible for the deb packaging to guarantee the use of
specific versions.  Especially "pillow" is problematic, since it gets
compiled and thus requires a set of libraries, that differs from the
ones available in our target release of Debian.

The upgrade of makefilet also disables automatic upgrades.