deb: add dependency packages for the supported database packages
By default the package "grouprise-db-postgresql" will be installed.
This ensures that all relevant dependencies for the given database
engine are installed.  Thus "pg_dump" (e.g. for database backup during
package upgrades) is guaranteed to be available (see #677), if the
selected backend package was installed alongside the main grouprise
package.

The following details are not yet implemented:
* no database configuration is done automatically
* the default grouprise configuration still refers to a PostgreSQL
  connection, even if "grouprise-db-sqlite" is installed

Closes: #677