Scores: evaluate the activity of the last 100 days (instead of nothing)
The previous default value for "SCORE_CONTENT_AGE" was zero.  Thus no
content at all was used for calculating the activity score of an entity.

Previously (before e0561144 and ce22b8d4) the default was 100 (days).
The change to zero was probably a mistake.
Now we use the 100 days threshold as the default activity time window
again.

Closes: #686