mail handling: implement processing of single emails via "dot-forward"
Implement a minimal interface for processing incoming mails via a
mechanism similar to "dot-forward".  Many mail servers support this
interface for mail delivery.

This introduces a new command `import_mail_message`.  It expects the
mail message in STDIN and evaluates either the `DTLINE` environment
setting or the `Delivered-To` or `To` header of the mail.
This command is a drop-in replacement of `processincomingmessage`, which
is provided by `django_mailbox`.  The latter will be removed from
grouprise (see #692).

The new configuration setting `COLLECTOR_MAILBOX_ADDRESS` can be used
for configuring a single-mailbox delivery configuration.  It is very
similar to `MAILBOX_DELIVERED_TO_EMAIL` (used by `django_mailbox`), but
its meaning differs slightly.  Thus an alternative setting name is used.

Minimal manual mail submission:

    cat mail.txt | DTLINE="Delivered-To: foo@example.org" grouprisectl import_mail_message

Example content of a `.forward` file:

    | grouprisectl import_mail_message