Remove "django_mailbox"
The django_mailbox integration provided the following features:
* receive mails via the "dotforward" mechanism
* retrieve mails from a remote IMAP mailbox

The "dotforward" mechanism is implemented directly now.
The retrieval of mails from a remote IMAP mailbox is not available anymore.

The preferred alternative for mail processing is grouprise's LMTP
server.  The retrieval of mails from an IMAP mailbox breaks the handling
of CC and BCC addresses, thus it is not recommended, anyway.

Closes: #692