Add groups to search index
Previously searching for the name (or description) of a group did not
return any results.

The visualization in the search result view is not optimal, since
articles use a gray background.  Additionally group results could be
shown separately on top of the search results, since they are probably
more relevant.