feat(matrix_chat): remove upload of room avatar
The previous implementation specified the room avatar via an https-URL
(pointing to the media published by grouprise).
But the most commonly used Matrix client (element) uses only "mxc://"
URLs.  These are based on file uploads to the Matrix server.
It would be rather hard to synchronize avatars due to this storage
layer, since we cannot determine, whether an avatar needs an update or
not.

Thus we do not support the synchronization of avatar images between the
grouprise group and the corresponding Matrix rooms for now.