feat(matrix_chat): remove references to django-cas-server
The [django-cas-server](https://github.com/nitmir/django-cas-server/)
works nicely as a single-sign-on session provider, but since its session
is not related to the grouprise session (in the same Django
application), it requires an additional login activity (see
[django-cas-server#70](https://github.com/nitmir/django-cas-server/issues/70)).

Thus we stick with the (local-only) grouprise authentication module for
matrix-synapse for now.
In the future we may want to explore the possibility of using an OIDC
provider module, which is tied to our grouprise session.