fix(settings): fix lazy-evaluation support
The previous implementation did work, since it was applied to the
instance (instead of the class).

Now we re-implement the functionality of "cached_property".

After the next Debian release (Bullseye), we could switch to
"functools.cached_property".  Then we will need to call `__set_name__`
afterwards in order to work around a design deficit of cached_property
in combination with `setattr`.
See: https://bugs.python.org/issue38517