build: use absolute filename for configuration
A few actions (e.g. `python3 manage.py compilemesages`) are executed
in a different directory (due to their limited configurability).
This causes problems with the discovery of grouprise's configuration
file (e.g. `grouprise.yaml` is not deterministic anymore).

In order to user the same configuration file for all actions,
we need to configure the configuration filename via the environment
variable `GROUPRISE_CONFIG`.