feat(settings): add temporary support for legacy configuration format
The existence of the previous configuration file
(/etc/grouprise/settings.py) is tested before loading settings.
In case this file exists, it will be loaded as a module.
A warning message is emitted in order to encourage migration to the new
format.

This should ensure that upgrades work without manual adjustments.