refactor(deb): remove triggered call of "collectstatic"
Previously upgrades of related python packages (e.g. django) triggered a
"collectstatic" action via dpkg.
But recently we enabled "collectstatic" during the start of the uwsgi
process, thus we can omit the triggered package operation.
This will remove a fragile part of the packaging, since it will break
package upgrades, if grouprise is not properly configured for some
reason.