feat(matrix_chat): resolve chat room URLs based on the web client of the user
Previously the chat room URLs resolved into links for the local instance
of the "element" web client.
Now the resolved link is based on the matrix ID of the requesting user
(if he is logged in).
The following steps are used:
1. extract the domain from the matrix ID of the user
2. determine the homeserver URL of the matrix domain
3. determine the announced web client URL, if available
   (via something like `https://example.org:8448/_matrix/client/`)

The default fallback in case of failures refers to the `matrix.to` web
service.
Users of the local matrix domain are redirected to `/stadt/chat/...` (as
before).

In the future we may want to use the new `matrix://` scheme instead,
since this would respect the local configuration of the user.