fix(events): improve handling of ongoing events
Previously ongoing events (started before *now*, ending after *now*)
were not displayed as "upcoming".
This is suitable for events lasting only an hour.
Events with longer duration (e.g. multiple days) should be shown in the
filtered lists as "upcoming".

Now events are shown in the "day" view, if they partially happen on that
day.
Events are shown in the "upcoming" event list, if they are either
ongoing or they start soon.  Maybe we will need to change this, if long
(multi-day) events are becoming more common and thus they would block
too many slots in the "upcoming" view.

Closes: #341
Closes: #409