fix(settings): emit clean error message without traceback while loading settings
Previously a stracktrace was emitted in case of a `ConfigError` exception
during the configuration loading procedure.
Now we handle this exception and reduce the output noise to the minimum.
This should improve readability.

The previous behaviour (raising `ConfigError`) instead of emitting the
error and exiting, can be restored by setting the environment variable
`GROUPRISE_EXIT_ON_CONFIG_ERROR` to `0`.