refactor(deb): move user handling from uwsgi to systemd/sysvinit
Previously we switched the user context via uwsgi.
This causes problems with creating directories and managing their
ownership (e.g. `RuntimeDirectory` in the service unit is owned by
root).
Moving the complete user handling to the service unit (or sysvinit)
removes this problem.
Now the uwsgi instance is executed in the context of the target user.