refactor: allow nesting of async function execution
Previously we used `asyncio.run` in a few place (e.g. in signal
handlers).
But this conflicts with the recently introduced cachalot, which relies
on async function calls for cache invalidation.
Thus signal handlers may be called within a running event loop.
We can detect this now and handle it safely.