build: simplify bumpversion configuration
Previously we used a regular expression for determining the proper
places for version replacements.
But for some mysterious reason, the regular expressions are not working
anymore (and according to bumpversion's source code they should have
never worked).
Weird :(

Let us just rely on the likely assumption, that only one occurrence of
the version string exists in each file.