fix(matrix_chat): proper redirection to Matrix chat rooms based on aliases
Previously room IDs were used as part of the Element URL returned to the
requester.  But these room IDs only work, if the user has already joined
the room.  All other users receive a confusing message:
 https://github.com/vector-im/element-web/issues/11526

Instead of using Matrix room IDs, we are supposed to use public room
names ("aliases").  The former are implementation details, which we
should not use in public:
 https://github.com/vector-im/element-web/issues/2925

Now we try to determine the public name of a room when returning the
redirect URL.