fix(setup_site): allow repeated execution of `setup_site` command
Previously the procedure failed to run, if called again (due to
duplicate articles being created).
Now it possible to change individual settings easily.
The current value for each prompted setting is used as a default value.
Thus it is possible to simply step through all settings and answer only
the relevant one.