feat: allow groups to define a single location for use in maps
This introduces the new and optional `geo` app, which adds a settings
panel for groups, that allows each group to define a single location for
use in maps.

The geo app requires support for geo coordinates in the database
backend, i.e. the PostGIS extension for PostgreSQL. Users of sqlite must
switch to spatialite.

Refs #92