feat(packaging): add sensible maintenance mode based on systemd service state
Grouprise is usually started through systemd so it makes sense to create
the maintenance mode files there. However, since we’ve switched to
systemd managing the uWSGI app server, it’s not a good place to disable
app traffic. Other solutions using the web server might be working and
should be implemented in the specific deployment.