feat(tags): add support for tag groups
This adds support for the visual grouping of tags.
For tags in the form of #<group>:<name> additional attributes will be
added to the DOM element. For instance #Color:Red results in the
following HTML:

	<a href="..."
	   class="tag"
	   data-tag-group-key="color"
	   data-tag-group-name="Color">
		<span>Red</span>
	</a>

By default, the CSS introduced with this change will still display the
tag mostly like it would have been in the past (#Color:Red), but
instances may override this default style.