fix(matrix_chat): allow huey task execution in "eager" mode
Previously the handling of recently saved "Content" instances relied on
the slightly delayed execution of the task handler.
This delay does not exist, if huey is using "eager" mode (which happens
in tests).
Now we wait for the corresponding Assocation (which is created after a
"Content" instance) and thus we do not require a delay anymore.