fix: improve filtering of bot-like accounts
The following tests are added:
* configurable exception for omitting users, which joined in the last 30
  days
* ignore special grouprise users
  (FEED_IMPORTER_GESTALT_ID and UNKNOWN_GESTALT_ID)

In addition we now consider all users to be bots, if they log in within
three seconds after account registration.
Based on the current set of users on stadtgestalten.org almost all
bot-like accounts showed a delay of less than a second between
registration and login.

See: #713