fix(matrix_commander): implement proper interpretation of login delay
Previously I assumed, that a short delay between registration and login
can be interpreted as a bot-like behavior (being automated).
But in reality, this is just the automatic behavior of grouprise: new
users are immediately logged in right after registration.

Thus the time delta between "last_login" and "date_joined" can only be
used for determining, whether a user logged in *again* after the
registration.

This reduces the number of potential bot accounts on stadtgestalten.org
significantly (from 8000 down to 500).
Probably the "can_login" test is misleading here (creating false
negatives).