fix(logrotate): proper switching between log files after rotating
Since we do not (can not?) signal uwsgi, that it should re-open the log
file, we use logrotate's "copytruncate" feature instead.
It ensures, that the old log *content* is moved to the next file, while
the current log file is truncated and keeps receiving new log data.

The uwsgi package for Debian uses the same approach for logrotate.