refactor: compatibility with Django's (v3.0 or later) partial async implementation
Django 3.0 started to implement partial support for async processing.
This partial implementation is in conflict with our previous usage of
async (especially when using the Matrix client library `nio`), since it
is not allowed anymore to use the ORM in an async context.

The following changes are applied:
* replace `run_async` (our custom implementation) with `async_to_sync`
* use the `async_to_sync` as a decorator for functions, which are
  expected to be sync, but rely heavily on async operations
  (e.g. a `db_task` for huey or `def handle` in management commands)
* use `await sync_to_async` whenever we access the ORM in an async
  context

The python package "asgiref" is required now.

Theoretically these changes should be compatible with Django 3.0 as well
as previous versions of Django.
But at least the tests fail with Django 2.x, 3.0 and 3.1 due to database
access locking errors (which cannot be fixed with `threading.Lock`).
Thus the minimum working Django version seems to be 3.2.