fix(matrix_chat): avoid access to lazily resolved settings in async context
The huey tasks are executed via the `run_huey` command.
But sadly Django's command running procedure does not guarantee, that
the `setup` function for each application is called.
Thus we need to ensure for each single huey task, that the settings are
resolved (and thus no ORM requests are triggered).