fix(deb): tolerate upgrade failure for element-web during package upgrade
Previously a package upgrade failed, if the local `element-web` instance
could not be updated somehow (e.g. due to a network failure).
But this update is only relevant during the very first package
installation (in order to configure the package properly).
Thus we ignore errors now, if everything was already configured once
before.