feat(matrix_chat): send matrix messages as `m.message` instead of `m.notice`
In general `m.notice`[1] would be more appropriate for bot messages, but
we assume that most users will regard the grouprise chat bot as a
deliberately choosen communication partner and thus we are allowed to
raise the priority of its messages.

In the future we may want to make this setting configurable in certain
contexts (e.g. in group settings).
But probably such technical details are of limited use for our target
audience.

[1] https://spec.matrix.org/latest/client-server-api/#mnotice