feat(matrix_chat): reduce default power level for public rooms to "standard"
Previously *everyone* in a public room had the same level ("moderator").
But this causes problems, if an external user joins the room and starts
to spam, since the group members (also having moderator rights) cannot
ban the external user, which has the same rights.

Now the default power level in a private room is reduced to "standard".
But every member of the group is raised to the "moderator" power level.

Closes: #778