feat: increase performance of filesystem-based cache for large sites
The performance of Django's builtin filesystem-based cache suffers for a
growing number of stored cache files (e.g. 14k in one instance).

The switch to the external implementation `diskcache` reduces the time
required for an update of the search index from 700 seconds down to 105
seconds.

Profiling data for the Django implementation revealed, that most of the
time was spent processing filename assembly and for going through the
flat directory storage again and again.

The `diskcache` package is now mentioned in the `Recommends` section of
the deb package.
It is automatically used, if the cache storage `filesystem` is
configured and if it is installed.
If it is unavailable, then Django's builtin implementation is used,
instead.

See: https://code.djangoproject.com/ticket/11260
See: https://code.djangoproject.com/ticket/29994
See: #765