tests: add `assertCreatedModelInstance` context
The context can be used for verifying, that a certain number of model
instances is created during a context.
In addition it provides access to these new instances.

This context be used for retrieving the new instances being created
by a `self.client.post` call (e.g. for creating articles, conversations
and other content).