feat(dx): detach webpack dev server listen address from public host
Sometimes we want the dev server to listen on a specific or all network
interfaces but expose it on another server. This is the case in docker
containers were we would want to bind the dev server to `0.0.0.0` but
map the port to our local host and the public host therefore becomes
`localhost`.