feat(dx): use separate directory for asset template
This is in preparation of our docker setup where we want to have a
shared named volume for our webpack dev server and the backend server.
Because we can’t mount subdirectories from named volumes we need to have
a separate directory for the asset template that is included by our
django application so that the mount doesn’t overlay other template
files in the directory.