fix(settings): revise config path order
In 782bb80a0c we’ve ordered the config paths so that intented paths come
first followed by paths with high privilege. This makes sense in a
production setup, because we don’t want to load config files from
directories that are writable by the system user running the backend
daemon. From a user/developer standpoint however, it is contrary to the
established convention that local configuration is applied on top or in
favor of system configuration.

Instead of defining a more or less arbitrary configuration file lookup
order, we now ensure that intent is always respected, so that if
grouprise is started with the GROUPRISE_CONFIG variable it will only
consider the specified directory/file. This way administrators and
container images can specifiy where the configuration will be placed.