fix(docker): await database connection in docker image
The first docker-compose run will almost always result in a broken
grouprise-backend service, because there is a race condition between the
database bootstrapping process and the attempt to run the initial
database migrate which require a working database connection.

The new database_await_connection simply awaits a working database
connection and only quits after a number of failed retries or in case of
a successful connection attempt.