feat: replace "simplemathcaptcha" with manually written trivial questions
Recently a bit of spam passed through the simplemathcaptcha, thus bots
seem to be able to adapt to them.
Now we use manually written trivial questions instead (similar to the
"Textchas" used in MoinMoin wiki).

The list of questions should surely be expanded quite a bit.
Maybe somewhen in the future we will need to turn this into site-local
questions in order to evade clever bots.