fix(matrix_chat): work around changed Matrix path for client discovery
When a user clicks on a chat room link, we reply with a redirect to the
user's web client.
Somewhen around matrix-synapse v1.50 the path `/_matrix/client` stopped
returning a redirect to the site-local web client.
Thus we maintain a manual list of Matrix web clients until either
MSC2312 [1] is supported or a suitable redirect is published again
(see [2]).

See #793

[1] https://github.com/matrix-org/matrix-spec-proposals/pull/2312
[2] https://github.com/matrix-org/synapse/issues/14722