fix(matrix_chat): remove matrix ID from a room even after a user is deleted
Previously the matrix ID of a recently removed user account was not
removed from its group's Matrix rooms, since the account object was already
gone, when the (delayed) task was executed.

Now all relevant information (Matrix ID of the user and the group rooms) are
retrieved *before* the delayed task function is called.