fix: avoid related-query with in-memory tag instance
Previously (before Django v4 or v5) it was possible to run a
related-query with in-memory instance (i.e. without saving a the tag).
But now we need to restrict our queries to existing objects.

Simply returning an empty queryset when retrieving content related to
a non-existing seems obvious.
Hard to say, why the original implementation tried to continue even if
the tag did not exist.