fix(dx): add custom stylesheets to body
This fixes a problem during development: styles added by the webpack dev
server will be added to the head but not add the position of the script
that includes the dev server files.

This may cause problems if the CSS selectors in the custom stylesheet
have the same specificity as the selectors in the default grouprise
theme. In this case the grouprise theme would take precedence during
development but not in production.