feat(geo): add support for restricted views in maps
The POI component now allows site admins to restrict the visible part of
the map. Currently only "points" is supported which restricts the view
to the points on the map. Additional options can be given as query
parameters.

Example:

{% include 'geo/poi.html' with api_loader_query='group'
restrict_view='points?pad=0.15&disableForPopups=true' %}

This displays a map rendering a marker for each group with an associated
location. The view is restricted to the set of displayed locations. The
bounding box of that restricted view is padded by 15%  (see:
https://leafletjs.com/reference.html#latlngbounds-pad). View restriction
is lifted when popups are opened so that they can fit on the map.